Fixed some broken coding conventions and some non-compiling classes
Fixed broken coding convention number 9 (No methods with over 5
parameters) in Game.java, ConnectFourGame.java and OthelloGame.java,
and changed the classes that were calling those methods, so that they
are calling the new method with less parameters.
Also changed a couple of lines in Player.java and HumanPlayer.java, as
the did not seem to be compiling.