C4 and Othello loading
So I've updated Othello to now load in the file and print the info to
terminal and I've begun working on C4 actually loading. As of right now
it will set the pieces on the board properly and the players names and
types however this is HARDCODED i.e. it's not doing it based off of the
saved data yet. I'll need to change this eventually. Also it's not
changing the names in the title bar or the time properly. Still pretty
neat though.