C4 loading
The time and player labels will now update when the game is loaded. The
label itself does take into account their name and colour now however
the game doesn't I'm pretty sure. Also been tinkering with Othello to
try and get it to the same point as C4 however they've implemented it so
differently it's taking a while.